fix bug in notebook used for SRA upload
There was a bug in the notebook (`SRA_upload/create_submission_sheet.ipynb`) that generates the list of FASTQ files to upload to the SRA. That bug is fixed in this commit, it is just the addition of a single `.strip()` to the filenames before globbing them. The bug only affected samples for which there were multiple different paths listed semi-colon delimited in the `data/barcode_runs.csv` file. For this repo, that is only the CB6 samples. The bug caused only files matching the first path to get uploaded, which in this particular case led to two FASTQ files to fail to be uploaded for each CB6 antibody.